Implementation instead of test abstract class
