--- vaadin-spring-security/src/test/java/com/vaadin/spring/access/SecuredViewAccessControl.java
@@ -1,47 +1,0 @@
-/*
- * Copyright 2015-2017 The original authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.vaadin.spring.access;
-
-import com.vaadin.spring.annotation.SpringComponent;
-import com.vaadin.ui.UI;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.context.SecurityContext;
-import org.springframework.security.core.context.SecurityContextHolder;
-
-import java.util.Collection;
-
-@SpringComponent
-public class SecuredViewAccessControl extends AbstractSecuredViewAccessControl {
-
-    @Override
-    protected boolean isAccessGranted(UI ui, String[] securityConfigurationAttributes) {
-        SecurityContext context = SecurityContextHolder.getContext();
-        Authentication authentication = context.getAuthentication();
-        if (authentication == null) {
-            return false;
-        }
-        Collection<? extends GrantedAuthority> authorities = authentication.getAuthorities();
-        for (GrantedAuthority authority : authorities) {
-            for (String attribute : securityConfigurationAttributes) {
-                if (authority.getAuthority().contains(attribute)) {
-                    return true;
-                }
-            }
-        }
-        return false;
-    }
-}
